WIP: introduce SafeTransactionBuilder
Please reword, split this commit up, extend it or change it as you see
fit. This is basically the flow for bundling safe transactions for the
listing side, which is not what we immediately need, as we're tackling
the buy side. You can keep this around for when we do implement
listings.